debugging

+++ VariableBase.Mathematics/VariableBase.Mathematics/Interfaces/IBasicMathAlgorithm.cs
@@ -7,24 +7,24 @@ namespace VariableBase.Mathematics.Interfaces
 {
     public interface IBasicMathAlgorithm
     {
-        IList<Decimal> AsSegments(Double rawDouble);
-        ReadOnlyCollection<Decimal> Add(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Tuple<ReadOnlyCollection<Decimal>, ReadOnlyCollection<Decimal>, ReadOnlyCollection<Decimal>> Divide(ReadOnlyCollection<Decimal> numerator, ReadOnlyCollection<Decimal> denominator, ReadOnlyCollection<Decimal> hint = null);
-        ReadOnlyCollection<Decimal> GetWholeNumberSomewhereBetween(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b, Double variance = 0);
-        Boolean IsBottom(ReadOnlyCollection<Decimal> number);
-        Boolean IsEqual(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsEven(ReadOnlyCollection<Decimal> a);
-        Boolean IsFirst(ReadOnlyCollection<Decimal> number);
-        Boolean IsGreaterThan(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsGreaterThanOrEqualTo(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsLessThan(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsLessThanOrEqualTo(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsNotEqual(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        Boolean IsOdd(ReadOnlyCollection<Decimal> a);
-        ReadOnlyCollection<Decimal> Multiply(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
-        ReadOnlyCollection<Decimal> PowerOfBase(ReadOnlyCollection<Decimal> a, Int32 times);
-        ReadOnlyCollection<Decimal> Square(ReadOnlyCollection<Decimal> a);
-        Tuple<ReadOnlyCollection<Decimal>, ReadOnlyCollection<Decimal>, ReadOnlyCollection<Decimal>> SquareRoot(ReadOnlyCollection<Decimal> number);
-        ReadOnlyCollection<Decimal> Subtract(ReadOnlyCollection<Decimal> a, ReadOnlyCollection<Decimal> b);
+        IList<Double> AsSegments(Double rawDouble);
+        ReadOnlyCollection<Double> Add(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Tuple<ReadOnlyCollection<Double>, ReadOnlyCollection<Double>, ReadOnlyCollection<Double>> Divide(ReadOnlyCollection<Double> numerator, ReadOnlyCollection<Double> denominator, ReadOnlyCollection<Double> hint = null);
+        ReadOnlyCollection<Double> GetWholeNumberSomewhereBetween(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b, Double variance = 0);
+        Boolean IsBottom(ReadOnlyCollection<Double> number);
+        Boolean IsEqual(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsEven(ReadOnlyCollection<Double> a);
+        Boolean IsFirst(ReadOnlyCollection<Double> number);
+        Boolean IsGreaterThan(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsGreaterThanOrEqualTo(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsLessThan(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsLessThanOrEqualTo(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsNotEqual(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        Boolean IsOdd(ReadOnlyCollection<Double> a);
+        ReadOnlyCollection<Double> Multiply(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
+        ReadOnlyCollection<Double> PowerOfBase(ReadOnlyCollection<Double> a, Int32 times);
+        ReadOnlyCollection<Double> Square(ReadOnlyCollection<Double> a);
+        Tuple<ReadOnlyCollection<Double>, ReadOnlyCollection<Double>, ReadOnlyCollection<Double>> SquareRoot(ReadOnlyCollection<Double> number);
+        ReadOnlyCollection<Double> Subtract(ReadOnlyCollection<Double> a, ReadOnlyCollection<Double> b);
     }
 }
